Changes to Royalty--

+++ Backend_UMR_Work_Program/DataModels/WP_OIL_CONDENSATE_PRODUCTION_ACTIVITIES_Contract_Type.cs
@@ -6,7 +6,7 @@ namespace Backend_UMR_Work_Program.Models
     public partial class WP_OIL_CONDENSATE_PRODUCTION_ACTIVITIES_Contract_Type
     {
         public string? Year_of_WP { get; set; }
-        public int? Total_Reconciled_National_Crude_Oil_Production { get; set; }
+        public decimal? Total_Reconciled_National_Crude_Oil_Production { get; set; }
         public string? Contract_Type { get; set; }
     }
 }
